pyproject.toml
[tool.poetry]
name = "strawberry-django-file-upload"
version = "0.1.0"
description = ""
authors = ["Marco Acierno <marcoaciernoemail@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.9"
Pillow = "^7.2.0"
strawberry-graphql = "^0.29.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
